First commit
This is the result of my long and painful attempt -as a non developer using AIs- to create a functional and understandable separate addon for each function from the complete script (>1000 lines) created by Marcin Zielinski on my demand (Gauthier Kervyn). The attached .blend files show the setup inside Blender when it is needed. It can surely be improved. I believe that 'Z-Hide' still doesn't work correctly.